オーダー生成、支払いまで完成、バグ修正

+++ src/main/java/com/saihou/mapper/OrderItemMapper.java
@@ -26,5 +26,5 @@ public interface OrderItemMapper {
 
     int update(OrderItem orderItem);
 
-    int delete(Long id);
+    int delete(Integer id);
 }
